Refactor: Product 가격 int타입으로 변경, User 연관관계에서 Payment 제거

+++ src/main/java/com/chone/server/domains/order/dto/response/CreateOrderResponse.java
@@ -16,7 +16,7 @@ public record CreateOrderResponse(
         order.getId(),
         new OrderResponse(
             order.getId(),
-            order.getTotalPrice().intValue(),
+            order.getTotalPrice(),
             order.getOrderType().getDescription(),
             order.getStatus().getDescription(),
             order.getCreatedAt()),
